opea-docker/asr/extra-requirements.txt
TTS
transformers
datasets
soundfile
librosa
scipy
pydantic
python-dotenv
requests
openai-whisper